+++ sdk-core/src/main/java/co/reachfive/identity/sdk/core/models/responses/webAuthn/RegistrationOptions.kt
@@ -4,6 +4,7 @@ import android.os.Parcelable
 import co.reachfive.identity.sdk.core.utils.WebAuthn
 import com.google.android.gms.fido.fido2.api.common.*
 import com.google.gson.annotations.SerializedName
+import kotlinx.parcelize.IgnoredOnParcel
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
@@ -12,6 +13,9 @@ data class RegistrationOptions(
     val friendlyName: String,
     val options: CredentialCreationOptions
 ) : Parcelable {
+    @IgnoredOnParcel
+    val publicKeyId: String = options.publicKey.rp.id
+
     fun toFido2Model(): PublicKeyCredentialCreationOptions {
         val publicKey = options.publicKey
